Moved cart service into own folder.
