+++ src/app/pages/view-all-products/view-all-products.component.ts
@@ -7,7 +7,7 @@ import {
   faSearch,
 } from '@fortawesome/free-solid-svg-icons';
 import { Subscription } from 'rxjs';
-import { CartService } from 'src/app/services/cart.service';
+import { CartService } from 'src/app/services/cart/cart.service';
 
 @Component({
   selector: 'app-view-all-products',
